+++ include/ignition/gazebo/components/JointAxis.hh
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *
-*/
-#ifndef IGNITION_GAZEBO_COMPONENTS_INERTIAL_HH_
-#define IGNITION_GAZEBO_COMPONENTS_INERTIAL_HH_
+ */
+#ifndef IGNITION_GAZEBO_COMPONENTS_JOINTAXIS_HH_
+#define IGNITION_GAZEBO_COMPONENTS_JOINTAXIS_HH_
 
-#include <ignition/math/Inertial.hh>
+#include <sdf/JointAxis.hh>
 
 #include <ignition/gazebo/config.hh>
 #include <ignition/gazebo/Export.hh>
@@ -28,15 +28,20 @@ namespace ignition
 {
 namespace gazebo
 {
-// Inline bracket to help doxygen filtering.
-inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
 namespace components
 {
-  /// \brief A component type that contains inertial, ignition::math::Inertiald,
-  /// information.
-  using Inertial = SimpleWrapper<ignition::math::Inertiald, class InertialTag>;
-}
+  // Inline bracket to help doxygen filtering.
+  inline namespace IGNITION_GAZEBO_VERSION_NAMESPACE {
+  /// \brief A component that contains the joint axis . This is a simple wrapper
+  /// around sdf::JointAxis
+  using JointAxis = SimpleWrapper<sdf::JointAxis, class JointAxisTag>;
+
+  /// \brief A component that contains the second joint axis for joints with two
+  /// axes. This is a simple wrapper around sdf::JointAxis
+  using JointAxis2 = SimpleWrapper<sdf::JointAxis, class JointAxis2Tag>;
+  }
 }
 }
 }
 #endif
+